+++ src/felleskomponenter/menypanel/menypunkter/familieforhold/medfolgendeFamilie/redigerer.less
@@ -1,4 +1,4 @@
-.medfolgende-barn__redigerer {
+.medfolgende-familie__redigerer {
   .slett-symbol {
     margin-top: 2.4em;
     display: inline-block;
