+++ src/felleskomponenter/menypanel/menypunkter/familieforhold/medfolgendeFamilie/redigeringUtfort.less
@@ -1,4 +1,4 @@
-.medfolgende-barn__redigeringutfort {
+.medfolgende-familie__redigeringutfort {
   .header {
     margin: 0.3em 0;
   }
