+++ src/felleskomponenter/menypanel/menypunkter/familieforhold/medfolgendeFamilie/redigeringUtfort.tsx
@@ -30,7 +30,7 @@ const Rad = ({ navn, idNummer }: RadProps) => (
 );
 
 const RedigeringUtfort = ({ verdier }: EnRedigeringsknappListeRedigeringUtfort<KV.Form.MedfolgendeBarn>) => (
-  <div className="medfolgende-barn__redigeringutfort">
+  <div className="medfolgende-familie__redigeringutfort">
     <ExpandableList
       elements={verdier}
       header={<Header />}
